extract CLI Ink components into src/cli/components with tests
