+++ src/cli/cli.tsx
@@ -2,8 +2,7 @@
 import path from "node:path";
 import { scheduler } from "node:timers/promises";
 import React, { useEffect, useRef, useState } from "react";
-import { Box, render, Text, useApp, useInput } from "ink";
-import { marked, type Token, type Tokens } from "marked";
+import { Box, render, useApp } from "ink";
 import {
   configureAuthProvider,
   listAuthProviderTools,
@@ -18,12 +17,9 @@ import {
 } from "../ingestion/code-mode.js";
 import {
   commandEmitsTelemetry,
-  helpContent,
-  isDevelopmentMode,
   parseCommand,
   shouldRunNonInteractively,
   type CliCommand,
-  type HelpRow,
   type OpenWikiRunMode,
 } from "./commands.js";
 import { resolveStartupCommand } from "./startup.js";
@@ -37,42 +33,25 @@ import {
   getErrorDiagnostics,
   type ErrorDiagnostic,
 } from "./diagnostics/error-diagnostics.js";
-import { sanitizeHeaderValue } from "./diagnostics/sanitize.js";
-import type {
-  ChatInputMenuState,
-  ChatInputState,
-  SecretInputMode,
-  SlashCommandOption,
-} from "./input/types.js";
-import {
-  applyRawInputValue,
-  deleteAtInputCursor,
-  deleteBeforeInputCursor,
-  isEscapeInput,
-  isRawBackspaceInput,
-  moveInputCursor,
-} from "./input/cursor.js";
-import {
-  getCurrentModelOptionIndex,
-  getCurrentProviderOptionIndex,
-  getModelMenuOptions,
-  isMenuDownInput,
-  isMenuUpInput,
-  moveMenuSelection,
-  parseSlashInput,
-  slashCommandOptions,
-  syncMenuStateForInput,
-} from "./input/menu.js";
-import { formatSecretInputSummary } from "./input/secret.js";
 import type { RunLogItem } from "./run-log/types.js";
 import { appendRunLogEvent } from "./run-log/reducer.js";
+import type { CompletedRun } from "./components/types.js";
+import { Header } from "./components/header.js";
+import { PromptBlock, StatusLine } from "./components/primitives.js";
 import {
-  formatCwd,
-  getDisplayModelId,
-  getSpinnerFrame,
-  isExitMessage,
-  truncateLogOutput,
-} from "./format.js";
+  FirstRunNotice,
+  renderFirstRunNoticeText,
+} from "./components/first-run-notice.js";
+import {
+  AuthFixPanel,
+  CredentialDiagnosticsPanel,
+  DryRunView,
+  ErrorDiagnosticsPanel,
+  HelpView,
+} from "./components/panels.js";
+import { IngestionSummary, RunView } from "./components/run-view.js";
+import { ChatHistory, ChatInput } from "./components/chat.js";
+import { getDisplayModelId, isExitMessage } from "./format.js";
 import {
   formatPowerScheduleStatus,
   formatScheduleHeader,
@@ -98,9 +77,7 @@ import {
 } from "../config/env.js";
 import { createOpenWikiThreadId, runOpenWikiAgent } from "../agent/index.js";
 import { installCrashGuard } from "../agent/crash-guard.js";
-import { formatChatGptAccountFromEnv } from "../agent/openai-chatgpt-oauth.js";
 import { getErrorMessage } from "../platform/diagnostics.js";
-import { stripHtmlTags } from "../platform/utils.js";
 import {
   type OpenWikiRunEvent,
   type OpenWikiRunResult,
@@ -123,28 +100,17 @@ import {
 import {
   getDefaultModelId,
   getMissingProviderEnvKey,
-  getProviderApiKeyEnvKey,
   getProviderCredentialHint,
   getProviderLabel,
   getProviderModelOptions,
-  getProviderProjectEnvKey,
-  isValidModelId,
-  normalizeModelId,
-  normalizeProvider,
   OPENWIKI_PROVIDER_ENV_KEY,
   OPENWIKI_MODEL_ID_ENV_KEY,
-  providerUsesAwsSdkCredentials,
   resolveConfiguredProvider,
-  SELECTABLE_OPENWIKI_PROVIDERS,
-  OPENWIKI_VERSION,
   type OpenWikiProvider,
 } from "../config/constants.js";
 import type { OpenWikiCommand, OpenWikiRunOptions } from "../agent/types.js";
 import {
   firstRunNoticePending,
-  FIRST_RUN_NOTICE_BODY,
-  FIRST_RUN_NOTICE_OPT_OUT,
-  FIRST_RUN_NOTICE_VERIFY,
   withRunTelemetry,
   type RunTelemetryContext,
 } from "../telemetry/index.js";
@@ -189,109 +155,9 @@ type RunState =
       authFix?: AuthFix;
     };
 
-type CompletedRun = {
-  id: number;
-  command: OpenWikiCommand;
-  credentialDiagnostics?: CredentialDiagnostic[];
-  log: RunLogItem[];
-  message: string | null;
-  result: OpenWikiRunResult;
-};
-
 type AppProps = {
   command: CliCommand;
 };
-
-const OPENWIKI_LOGO_LINES = [
-  "  ___                  __        ___ _    _ ",
-  " / _ \\ _ __   ___ _ __ \\ \\      / (_) | _(_)",
-  "| | | | '_ \\ / _ \\ '_ \\ \\ \\ /\\ / /| | |/ / |",
-  "| |_| | |_) |  __/ | | | \\ V  V / | |   <| |",
-  " \\___/| .__/ \\___|_| |_|  \\_/\\_/  |_|_|\\_\\_|",
-  "      |_|",
-];
-const OPENWIKI_LOGO_WIDTH = Math.max(
-  ...OPENWIKI_LOGO_LINES.map((line) => line.length),
-);
-
-/** Frame/wrap width for the plain-text (print/non-TTY) first-run disclosure. */
-const FIRST_RUN_NOTICE_WIDTH = 64;
-
-/** Greedy word-wrap to `width` columns. Input carries no existing newlines. */
-function wrapText(text: string, width: number): string[] {
-  const lines: string[] = [];
-  let line = "";
-
-  for (const word of text.split(/\s+/)) {
-    if (line.length === 0) {
-      line = word;
-    } else if (line.length + 1 + word.length <= width) {
-      line += ` ${word}`;
-    } else {
-      lines.push(line);
-      line = word;
-    }
-  }
-  if (line.length > 0) {
-    lines.push(line);
-  }
-
-  return lines;
-}
-
-/**
- * The plain-text first-run disclosure for print/non-TTY output: the same copy as
- * the interactive box (single-sourced in telemetry/config.ts), framed with light
- * rules and wrapped to a fixed width. Rendered gray when stderr is a TTY, plain
- * when redirected so a captured log stays free of escape codes.
- */
-function renderFirstRunNoticeText(color: boolean): string {
-  const label = "OpenWiki telemetry";
-  const width = FIRST_RUN_NOTICE_WIDTH;
-  const topRule = `─── ${label} ${"─".repeat(Math.max(3, width - label.length - 5))}`;
-  const block = [
-    "",
-    topRule,
-    "",
-    ...wrapText(FIRST_RUN_NOTICE_BODY, width),
-    "",
-    ...wrapText(FIRST_RUN_NOTICE_OPT_OUT, width),
-    "",
-    ...wrapText(FIRST_RUN_NOTICE_VERIFY, width),
-    "─".repeat(width),
-    "",
-  ].join("\n");
-
-  return color ? `\u001b[90m${block}\u001b[39m` : block;
-}
-
-/**
- * The one-time telemetry disclosure, rendered as a box so it sits inline with
- * the rest of the TUI (mirrors SetupHeader's rounded style). The copy is
- * single-sourced in telemetry/config.ts; the print/non-TTY path renders the
- * same wording as plain text via renderFirstRunNoticeText.
- */
-function FirstRunNotice() {
-  return (
-    <Box
-      borderStyle="round"
-      borderColor="cyan"
-      flexDirection="column"
-      marginBottom={1}
-      paddingX={1}
-    >
-      <Text>
-        <Text bold color="cyan">
-          OpenWiki
-        </Text>{" "}
-        <Text color="gray">telemetry</Text>
-      </Text>
-      <Text color="white">{FIRST_RUN_NOTICE_BODY}</Text>
-      <Text color="white">{FIRST_RUN_NOTICE_OPT_OUT}</Text>
-      <Text color="white">{FIRST_RUN_NOTICE_VERIFY}</Text>
-    </Box>
-  );
-}
 
 function App({ command }: AppProps) {
   const app = useApp();
@@ -1067,1368 +933,6 @@ function App({ command }: AppProps) {
   );
 }
 
-function HelpView() {
-  return (
-    <Box flexDirection="column">
-      <Header modelId={null} subtitle={helpContent.description} />
-
-      <Panel title="Usage">
-        {helpContent.usage.map((line) => (
-          <Text key={line}> {line}</Text>
-        ))}
-      </Panel>
-
-      <Panel title="Commands">
-        <Rows rows={helpContent.commands} />
-      </Panel>
-
-      <Panel title="Options">
-        <Rows rows={helpContent.options} />
-      </Panel>
-
-      {isDevelopmentMode() ? (
-        <Panel title="Development Options">
-          <Rows rows={helpContent.developmentOptions} />
-        </Panel>
-      ) : null}
-
-      <Panel title="Examples">
-        {helpContent.examples.map((line) => (
-          <Text key={line}> {line}</Text>
-        ))}
-        {isDevelopmentMode()
-          ? helpContent.developmentExamples.map((line) => (
-              <Text key={line}> {line}</Text>
-            ))
-          : null}
-      </Panel>
-    </Box>
-  );
-}
-
-function DryRunView({
-  command,
-  modelId,
-  shouldStart,
-  userMessage,
-}: {
-  command: OpenWikiCommand;
-  modelId: string | null;
-  shouldStart: boolean;
-  userMessage: string | null;
-}) {
-  return (
-    <Box flexDirection="column">
-      <Header modelId={modelId} subtitle="Development dry run" />
-      <Panel title="Execution Plan">
-        <StatusLine
-          tone="active"
-          label="Command"
-          value={`openwiki ${command}`}
-        />
-        <StatusLine tone="muted" label="Mode" value={command} />
-        <StatusLine
-          tone="muted"
-          label="Credentials"
-          value="not read or requested"
-        />
-        <StatusLine
-          tone="muted"
-          label="Model"
-          value={
-            modelId ??
-            `saved setting or ${getDefaultModelId(resolveConfiguredProvider())}`
-          }
-        />
-        <StatusLine tone="muted" label="Agent" value="not invoked" />
-        <StatusLine tone="muted" label="Writes" value="no files or metadata" />
-        <StatusLine tone="muted" label="Output" value="~/.openwiki/wiki" />
-        <StatusLine
-          tone="muted"
-          label="Startup"
-          value={shouldStart ? "would start run" : "would open chat"}
-        />
-        {userMessage ? (
-          <StatusLine tone="muted" label="Message" value={userMessage} />
-        ) : null}
-      </Panel>
-    </Box>
-  );
-}
-
-function CredentialDiagnosticsPanel({
-  diagnostics,
-}: {
-  diagnostics: CredentialDiagnostic[];
-}) {
-  return (
-    <Panel title="Credential Diagnostics">
-      <Text color="gray">Raw secret values are intentionally not printed.</Text>
-      {diagnostics.map((diagnostic) => (
-        <Box flexDirection="column" key={diagnostic.key} marginTop={1}>
-          <Text>
-            <Text bold>{diagnostic.key}</Text>{" "}
-            <Text color="gray">source={diagnostic.source}</Text>
-          </Text>
-          <Text>
-            length={diagnostic.length ?? "unset"} preview={diagnostic.preview}
-          </Text>
-          <Text color={diagnostic.warnings.length > 0 ? "yellow" : "gray"}>
-            warnings=
-            {diagnostic.warnings.length > 0
-              ? diagnostic.warnings.join(", ")
-              : "none"}
-          </Text>
-        </Box>
-      ))}
-    </Panel>
-  );
-}
-
-/**
- * The ordered "how to fix" steps for an auth failure. Shared by the interactive
- * panel and the --print stderr path so they stay in sync. Names env keys only,
- * never secret values.
- */
-function AuthFixPanel({ authFix }: { authFix: AuthFix }) {
-  const steps = getAuthFixSteps(authFix);
-
-  return (
-    <Panel title="How to fix">
-      <Text>Your provider rejected the credentials for this run.</Text>
-      {steps.map((step, index) => (
-        <Text key={step}>
-          <Text color="cyan">{index + 1}. </Text>
-          {step}
-        </Text>
-      ))}
-      <Text color="gray">For full detail, re-run with --debug.</Text>
-    </Panel>
-  );
-}
-
-function ErrorDiagnosticsPanel({
-  diagnostics,
-}: {
-  diagnostics: ErrorDiagnostic[];
-}) {
-  return (
-    <Panel title="Error Diagnostics">
-      <Text color="gray">
-        OPENWIKI_DEBUG=1 is enabled. Only allowlisted, non-secret error fields
-        are shown.
-      </Text>
-      {diagnostics.map((diagnostic) => (
-        <Text key={diagnostic.label}>
-          <Text bold>{diagnostic.label}</Text> {diagnostic.value}
-        </Text>
-      ))}
-    </Panel>
-  );
-}
-
-function Header({
-  compact = false,
-  modelId,
-  showLogo = true,
-  subtitle,
-}: {
-  compact?: boolean;
-  modelId?: string | null;
-  showLogo?: boolean;
-  subtitle: string;
-}) {
-  const terminalColumns = process.stdout.columns ?? 80;
-  const displayModelId = sanitizeHeaderValue(
-    modelId ??
-      process.env[OPENWIKI_MODEL_ID_ENV_KEY] ??
-      getDefaultModelId(resolveConfiguredProvider()),
-    Math.max(8, terminalColumns - 12),
-  );
-  const configuredProvider = resolveConfiguredProvider();
-  const displayProvider = getProviderLabel(configuredProvider);
-  const chatGptAccount =
-    configuredProvider === "openai-chatgpt"
-      ? formatChatGptAccountFromEnv()
-      : null;
-  const displayDirectory = sanitizeHeaderValue(
-    formatCwd(process.cwd()),
-    Math.max(8, terminalColumns - 17),
-  );
-  const shouldShowLogo = showLogo && terminalColumns > OPENWIKI_LOGO_WIDTH;
-  const tracingEnabled =
-    process.env.LANGCHAIN_TRACING_V2 === "true" &&
-    Boolean(process.env.LANGSMITH_API_KEY);
-
-  if (compact) {
-    return (
-      <Box flexDirection="column" marginBottom={1}>
-        <Text wrap="truncate">
-          <Text color="cyan">{">_ "}</Text>
-          <Text bold>OpenWiki</Text>{" "}
-          <Text color="gray">v{OPENWIKI_VERSION}</Text>{" "}
-          <Text color="gray">provider: </Text>
-          <Text color="white">{displayProvider}</Text>{" "}
-          {chatGptAccount ? (
-            <>
-              <Text color="gray">account: </Text>
-              <Text color="white">{chatGptAccount}</Text>{" "}
-            </>
-          ) : null}
-          <Text color="gray">model: </Text>
-          <Text color="white">{displayModelId}</Text>
-        </Text>
-        <Text>
-          <Text color={tracingEnabled ? "green" : "gray"}>
-            {tracingEnabled ? "* " : "- "}
-          </Text>
-          <Text color={tracingEnabled ? "green" : "gray"}>
-            LangSmith tracing {tracingEnabled ? "enabled" : "disabled"}
-          </Text>
-          <Text color="gray"> - </Text>
-          <Text color="cyan">{subtitle}</Text>
-        </Text>
-      </Box>
-    );
-  }
-
-  return (
-    <Box flexDirection="column" marginBottom={1}>
-      {shouldShowLogo ? (
-        <Box flexDirection="column" marginBottom={1}>
-          {OPENWIKI_LOGO_LINES.map((line) => (
-            <Text bold color="cyan" key={line} wrap="truncate">
-              {line}
-            </Text>
-          ))}
-        </Box>
-      ) : null}
-      <Box
-        borderColor="cyan"
-        borderStyle="round"
-        flexDirection="column"
-        marginBottom={1}
-        paddingX={1}
-      >
-        <Text>
-          <Text color="cyan">{">_ "}</Text>
-          <Text bold>OpenWiki</Text>{" "}
-          <Text color="gray">v{OPENWIKI_VERSION}</Text>{" "}
-          <Text color="gray">agent docs for codebases</Text>
-        </Text>
-        <Text>
-          <Text color="gray">provider: </Text>
-          <Text color="white">{displayProvider}</Text>
-        </Text>
-        {chatGptAccount ? (
-          <Text>
-            <Text color="gray">account: </Text>
-            <Text color="white">{chatGptAccount}</Text>
-          </Text>
-        ) : null}
-        <Text>
-          <Text color="gray">model: </Text>
-          <Text color="white">{displayModelId}</Text>
-        </Text>
-        <Text>
-          <Text color="gray">directory: </Text>
-          <Text color="white">{displayDirectory}</Text>
-        </Text>
-      </Box>
-      <Text>
-        <Text color={tracingEnabled ? "green" : "gray"}>
-          {tracingEnabled ? "* " : "- "}
-        </Text>
-        <Text color={tracingEnabled ? "green" : "gray"}>
-          LangSmith tracing {tracingEnabled ? "enabled" : "disabled"}
-        </Text>
-        <Text color="gray"> - </Text>
-        <Text color="cyan">{subtitle}</Text>
-      </Text>
-      <Text color="gray">
-        Tip: ask for a docs change, or use /exit when you are done.
-      </Text>
-    </Box>
-  );
-}
-
-type StatusLineProps = {
-  tone: "active" | "error" | "muted" | "success";
-  label: string;
-  value: string;
-};
-
-function StatusLine({ tone, label, value }: StatusLineProps) {
-  const color =
-    tone === "success"
-      ? "green"
-      : tone === "error"
-        ? "red"
-        : tone === "active"
-          ? "yellow"
-          : "gray";
-
-  return (
-    <Text>
-      <Text color={color}>* </Text>
-      <Text bold color={color}>
-        {label}
-      </Text>{" "}
-      <Text color={tone === "muted" ? "gray" : undefined}>{value}</Text>
-    </Text>
-  );
-}
-
-function IngestionSummary({ result }: { result: OpenWikiIngestionResult }) {
-  return (
-    <Panel title="Source Runs">
-      {result.results.map((sourceResult) => (
-        <StatusLine
-          key={sourceResult.sourceInstanceId}
-          label={sourceResult.displayName}
-          tone={sourceResult.status === "error" ? "error" : "success"}
-          value={`${sourceResult.status}; ${sourceResult.rawFiles.length} raw file(s)`}
-        />
-      ))}
-    </Panel>
-  );
-}
-
-type RunViewProps = {
-  command: OpenWikiCommand;
-  credentialDiagnostics?: CredentialDiagnostic[];
-  log: RunLogItem[];
-  done?: boolean;
-  message?: string | null;
-  modelId?: string | null;
-};
-
-function RunView({
-  command,
-  credentialDiagnostics,
-  log,
-  done = false,
-  message = null,
-  modelId = null,
-}: RunViewProps) {
-  const [animationFrame, setAnimationFrame] = useState(0);
-  const activeRunningToolId = getActiveRunningToolLogId(log);
-  const hasRunningTool = activeRunningToolId !== null;
-
-  useEffect(() => {
-    if (done || !hasRunningTool) {
-      return;
-    }
-
-    const interval = setInterval(() => {
-      setAnimationFrame((frame) => frame + 1);
-    }, 140);
-
-    return () => {
-      clearInterval(interval);
-    };
-  }, [done, hasRunningTool]);
-
-  return (
-    <Box flexDirection="column">
-      <Header
-        compact
-        modelId={modelId}
-        showLogo={false}
-        subtitle={done ? "Run complete" : "Agent running"}
-      />
-      {message ? <PromptBlock message={message} /> : null}
-      <Box flexDirection="column" marginBottom={1}>
-        <Text>
-          <Text color={done ? "green" : "cyan"}>* </Text>
-          <Text bold>{done ? "Complete" : "Working"}</Text>{" "}
-          <Text color="gray">openwiki {command}</Text>
-          {!done ? <Text color="gray"> - streaming</Text> : null}
-        </Text>
-        <Box flexDirection="column" marginLeft={2} marginTop={1}>
-          {log.length > 0 ? (
-            log.map((item) => (
-              <RunLogLine
-                activeRunningToolId={activeRunningToolId}
-                animationFrame={animationFrame}
-                item={item}
-                key={item.id}
-              />
-            ))
-          ) : (
-            <Text color="gray">Waiting for model output...</Text>
-          )}
-        </Box>
-      </Box>
-      {credentialDiagnostics ? (
-        <CredentialDiagnosticsPanel diagnostics={credentialDiagnostics} />
-      ) : null}
-    </Box>
-  );
-}
-
-function RunLogLine({
-  activeRunningToolId = null,
-  animationFrame = 0,
-  item,
-}: {
-  activeRunningToolId?: number | null;
-  animationFrame?: number;
-  item: RunLogItem;
-}) {
-  if (item.type === "tool") {
-    if (item.status === "running") {
-      const isActive = item.id === activeRunningToolId;
-
-      return (
-        <Box flexDirection="column" marginBottom={1}>
-          <Text>
-            <Text color={isActive ? "cyan" : "gray"}>
-              {isActive ? `${getSpinnerFrame(animationFrame)} ` : "* "}
-            </Text>
-            <Text bold={isActive} color={isActive ? "cyan" : "gray"}>
-              {item.content}
-            </Text>
-          </Text>
-          {isActive && item.call ? (
-            <Text color="gray"> {truncateLogOutput(item.call, "")}</Text>
-          ) : null}
-        </Box>
-      );
-    }
-
-    if (item.status === "error") {
-      return (
-        <Box flexDirection="column" marginBottom={1}>
-          <Text>
-            <Text bold color="red">
-              {"!! "}
-            </Text>
-            <Text bold color="red">
-              {item.content}
-            </Text>
-          </Text>
-        </Box>
-      );
-    }
-
-    return (
-      <Box flexDirection="column" marginBottom={1}>
-        <Text>
-          <Text color="green">{"* "}</Text>
-          <Text color="gray">{item.content}</Text>
-        </Text>
-      </Box>
-    );
-  }
-
-  if (item.type === "debug") {
-    return (
-      <Text>
-        <Text color="gray">- </Text>
-        <Text color="gray">{item.content}</Text>
-      </Text>
-    );
-  }
-
-  return (
-    <Box flexDirection="row">
-      <Text color="white">* </Text>
-      <Box flexDirection="column">
-        <MarkdownText markdown={item.content.trim()} />
-      </Box>
-    </Box>
-  );
-}
-
-function getActiveRunningToolLogId(log: RunLogItem[]): number | null {
-  for (let index = log.length - 1; index >= 0; index -= 1) {
-    const item = log[index];
-
-    if (item.type === "tool" && item.status === "running") {
-      return item.id;
-    }
-  }
-
-  return null;
-}
-
-function MarkdownText({ markdown }: { markdown: string }) {
-  const tokens = marked.lexer(markdown, {
-    async: false,
-    gfm: true,
-  });
-
-  return (
-    <Box flexDirection="column">
-      {tokens.map((token, index) => (
-        <MarkdownBlock
-          index={index}
-          key={`${token.type}-${index}`}
-          token={token}
-        />
-      ))}
-    </Box>
-  );
-}
-
-function MarkdownBlock({ index, token }: { index: number; token: Token }) {
-  if (token.type === "space" || token.type === "def" || token.type === "hr") {
-    return null;
-  }
-
-  if (token.type === "paragraph") {
-    return (
-      <Text wrap="wrap">
-        <InlineMarkdown tokens={getTokenChildren(token)} />
-      </Text>
-    );
-  }
-
-  if (token.type === "heading") {
-    return (
-      <Text wrap="wrap">
-        <InlineMarkdown tokens={getTokenChildren(token)} />
-      </Text>
-    );
-  }
-
-  if (token.type === "list") {
-    return (
-      <Box flexDirection="column">
-        {(token as Tokens.List).items.map((item, itemIndex) => (
-          <Text key={`${index}-${itemIndex}`} wrap="wrap">
-            <Text color="gray">
-              {(token as Tokens.List).ordered
-                ? `${Number((token as Tokens.List).start || 1) + itemIndex}. `
-                : "- "}
-            </Text>
-            <InlineMarkdown tokens={getTokenChildren(item)} />
-          </Text>
-        ))}
-      </Box>
-    );
-  }
-
-  if (token.type === "code") {
-    return <Text color="gray">{token.text}</Text>;
-  }
-
-  if (token.type === "blockquote") {
-    return (
-      <Text wrap="wrap">
-        <Text color="gray">| </Text>
-        <InlineMarkdown tokens={getTokenChildren(token)} />
-      </Text>
-    );
-  }
-
-  if (token.type === "table") {
-    return <Text color="gray">{renderPlainTable(token as Tokens.Table)}</Text>;
-  }
-
-  if (token.type === "html") {
-    return <Text wrap="wrap">{renderHtmlToken(token)}</Text>;
-  }
-
-  if (token.type === "text") {
-    return (
-      <Text wrap="wrap">
-        <InlineMarkdown tokens={token.tokens ?? [token]} />
-      </Text>
-    );
-  }
-
-  return <Text wrap="wrap">{token.raw}</Text>;
-}
-
-function InlineMarkdown({ tokens }: { tokens: Token[] }) {
-  return (
-    <>
-      {tokens.map((token, index) => (
-        <InlineMarkdownToken key={`${token.type}-${index}`} token={token} />
-      ))}
-    </>
-  );
-}
-
-function InlineMarkdownToken({ token }: { token: Token }) {
-  if (token.type === "text" || token.type === "escape") {
-    return <>{token.text}</>;
-  }
-
-  if (token.type === "strong") {
-    return (
-      <Text bold>
-        <InlineMarkdown tokens={getTokenChildren(token)} />
-      </Text>
-    );
-  }
-
-  if (token.type === "em") {
-    return (
-      <Text italic>
-        <InlineMarkdown tokens={getTokenChildren(token)} />
-      </Text>
-    );
-  }
-
-  if (token.type === "link") {
-    return (
-      <Text underline>
-        <InlineMarkdown tokens={getTokenChildren(token)} />
-      </Text>
-    );
-  }
-
-  if (token.type === "codespan") {
-    return <Text color="gray">{token.text}</Text>;
-  }
-
-  if (token.type === "br") {
-    return <>{"\n"}</>;
-  }
-
-  if (token.type === "del") {
-    return (
-      <Text strikethrough>
-        <InlineMarkdown tokens={getTokenChildren(token)} />
-      </Text>
-    );
-  }
-
-  if (token.type === "html") {
-    return <>{renderHtmlToken(token)}</>;
-  }
-
-  if ("tokens" in token && Array.isArray(token.tokens)) {
-    return <InlineMarkdown tokens={token.tokens} />;
-  }
-
-  return <>{token.raw}</>;
-}
-
-function getTokenChildren(token: Token): Token[] {
-  return "tokens" in token && Array.isArray(token.tokens) ? token.tokens : [];
-}
-
-function renderPlainTable(token: Tokens.Table): string {
-  const header = token.header.map((cell) => cell.text).join(" | ");
-  const rows = token.rows.map((row) =>
-    row.map((cell) => cell.text).join(" | "),
-  );
-
-  return [header, ...rows].filter(Boolean).join("\n");
-}
-
-function renderHtmlToken(token: Token): React.ReactNode {
-  const text =
-    "text" in token && typeof token.text === "string" ? token.text : token.raw;
-  const underlineMatch = text.match(/^<u>(.*)<\/u>$/isu);
-
-  if (underlineMatch) {
-    return <Text underline>{underlineMatch[1]}</Text>;
-  }
-
-  return stripHtmlTags(text);
-}
-
-function ChatHistory({ runs }: { runs: CompletedRun[] }) {
-  if (runs.length === 0) {
-    return null;
-  }
-
-  return (
-    <Box flexDirection="column">
-      {runs.map((run) => (
-        <Box flexDirection="column" key={run.id} marginBottom={1}>
-          {run.message ? <PromptBlock message={run.message} /> : null}
-          <Text>
-            <Text color="green">* </Text>
-            <Text bold>Complete</Text>{" "}
-            <Text color="gray">
-              openwiki {run.command} - {run.result.model}
-            </Text>
-          </Text>
-          <Box flexDirection="column" marginLeft={2} marginTop={1}>
-            {run.log.length > 0 ? (
-              run.log.map((item) => <RunLogLine item={item} key={item.id} />)
-            ) : (
-              <Text color="gray">No assistant output captured.</Text>
-            )}
-          </Box>
-        </Box>
-      ))}
-    </Box>
-  );
-}
-
-type ChatInputProps = {
-  currentModelId: string;
-  currentProvider: OpenWikiProvider;
-  onClear: () => void;
-  onCommandRun: (
-    command: Extract<OpenWikiCommand, "init" | "update">,
-    message: string | null,
-  ) => void;
-  onModelSelect: (modelId: string) => Promise<void>;
-  onProviderSelect: (provider: OpenWikiProvider) => Promise<void>;
-  onSubmit: (message: string) => void;
-};
-
-function ChatInput({
-  currentModelId,
-  currentProvider,
-  onClear,
-  onCommandRun,
-  onModelSelect,
-  onProviderSelect,
-  onSubmit,
-}: ChatInputProps) {
-  const [inputState, setInputState] = useState<ChatInputState>({
-    cursorPosition: 0,
-    value: "",
-  });
-  const [menuState, setMenuState] = useState<ChatInputMenuState>({
-    kind: "none",
-  });
-  const [error, setError] = useState<string | null>(null);
-  const [notice, setNotice] = useState<string | null>(null);
-  const [isSaving, setIsSaving] = useState(false);
-  const [secretInputMode, setSecretInputMode] =
-    useState<SecretInputMode | null>(null);
-  const input = inputState.value;
-  const cursorPosition = inputState.cursorPosition;
-
-  useEffect(() => {
-    if (secretInputMode !== null) {
-      return;
-    }
-
-    setMenuState((currentState) =>
-      syncMenuStateForInput(
-        input,
-        currentState,
-        currentModelId,
-        currentProvider,
-      ),
-    );
-  }, [currentModelId, currentProvider, input, secretInputMode]);
-
-  useInput((inputValue, key) => {
-    if (isSaving) {
-      return;
-    }
-
-    if (secretInputMode !== null) {
-      if (isEscapeInput(inputValue, key)) {
-        resetInput();
-        setSecretInputMode(null);
-        setNotice("Credential update canceled.");
-        return;
-      }
-
-      if (key.return) {
-        void saveSecretInput();
-        return;
-      }
-
-      if (key.backspace || isRawBackspaceInput(inputValue)) {
-        setInputState(deleteBeforeInputCursor);
-        return;
-      }
-
-      if (key.delete) {
-        setInputState(
-          inputValue.length === 0
-            ? deleteBeforeInputCursor
-            : deleteAtInputCursor,
-        );
-        return;
-      }
-
-      if (inputValue && !key.ctrl && !key.meta) {
-        setError(null);
-        setNotice(null);
-        setInputState((state) => applyRawInputValue(state, inputValue));
-      }
-
-      return;
-    }
-
-    if (isMenuUpInput(inputValue, key) && menuState.kind !== "none") {
-      setMenuState((state) =>
-        moveMenuSelection(state, -1, currentModelId, currentProvider),
-      );
-      return;
-    }
-
-    if (isMenuDownInput(inputValue, key) && menuState.kind !== "none") {
-      setMenuState((state) =>
-        moveMenuSelection(state, 1, currentModelId, currentProvider),
-      );
-      return;
-    }
-
-    if (key.return) {
-      void submitInput();
-      return;
-    }
-
-    if (isEscapeInput(inputValue, key) && menuState.kind !== "none") {
-      resetInput();
-      return;
-    }
-
-    if (key.leftArrow) {
-      setInputState((state) => moveInputCursor(state, -1));
-      return;
-    }
-
-    if (key.rightArrow) {
-      setInputState((state) => moveInputCursor(state, 1));
-      return;
-    }
-
-    if ((key.ctrl && inputValue === "a") || inputValue === "\u0001") {
-      setInputState((state) => ({
-        ...state,
-        cursorPosition: 0,
-      }));
-      return;
-    }
-
-    if ((key.ctrl && inputValue === "e") || inputValue === "\u0005") {
-      setInputState((state) => ({
-        ...state,
-        cursorPosition: state.value.length,
-      }));
-      return;
-    }
-
-    if (key.backspace || isRawBackspaceInput(inputValue)) {
-      setInputState(deleteBeforeInputCursor);
-      return;
-    }
-
-    if (key.delete) {
-      setInputState(
-        inputValue.length === 0 ? deleteBeforeInputCursor : deleteAtInputCursor,
-      );
-      return;
-    }
-
-    if (inputValue && !key.ctrl && !key.meta) {
-      setError(null);
-      setNotice(null);
-      setInputState((state) => applyRawInputValue(state, inputValue));
-    }
-  });
-
-  async function submitInput() {
-    const message = input.trim();
-
-    if (message.length === 0) {
-      setError("Enter a follow-up message.");
-      return;
-    }
-
-    if (message.startsWith("/")) {
-      await submitSlashInput(message);
-      return;
-    }
-
-    resetInput();
-    onSubmit(message);
-  }
-
-  async function submitSlashInput(message: string) {
-    if (message === "/" && menuState.kind === "commands") {
-      await runSlashCommand(slashCommandOptions[menuState.selectedIndex]);
-      return;
-    }
-
-    if (message === "/model" && menuState.kind === "model") {
-      await selectModelMenuOption(menuState.selectedIndex);
-      return;
-    }
-
-    if (message === "/provider" && menuState.kind === "provider") {
-      await selectProviderMenuOption(menuState.selectedIndex);
-      return;
-    }
-
-    const parsedCommand = parseSlashInput(message);
-
-    if (parsedCommand === null) {
-      setError(`Unknown command: ${message}`);
-      return;
-    }
-
-    await runSlashCommand(
-      parsedCommand.option,
-      parsedCommand.args.length > 0 ? parsedCommand.args : null,
-    );
-  }
-
-  async function runSlashCommand(
-    option: SlashCommandOption | undefined,
-    args: string | null = null,
-  ) {
-    if (!option) {
-      setError("Select a slash command.");
-      return;
-    }
-
-    if (option.id === "model") {
-      if (args && args.length > 0) {
-        await saveModelSelection(args);
-        return;
-      }
-
-      setError(null);
-      setNotice("Choose a model, or type /model <model-id>.");
-      setInputValue("/model");
-      setMenuState({
-        kind: "model",
-        selectedIndex: getCurrentModelOptionIndex(
-          currentModelId,
-          currentProvider,
-        ),
-      });
-      return;
-    }
-
-    if (option.id === "provider") {
-      if (args && args.length > 0) {
-        await saveProviderSelection(args);
-        return;
-      }
-
-      setError(null);
-      setNotice("Choose a provider, or type /provider <provider-id>.");
-      setInputValue("/provider");
-      setMenuState({
-        kind: "provider",
-        selectedIndex: getCurrentProviderOptionIndex(currentProvider),
-      });
-      return;
-    }
-
-    if (option.id === "api-key") {
-      if (args && args.length > 0) {
-        setError(
-          "Use the masked prompt for API keys; do not pass keys inline.",
-        );
-        return;
-      }
-
-      if (providerUsesAwsSdkCredentials(currentProvider)) {
-        setError(
-          `${getProviderLabel(currentProvider)} uses the AWS SDK credential chain; /api-key cannot safely configure an access-key pair. ${getProviderCredentialHint(currentProvider) ?? ""} Legacy BEDROCK_AWS_ACCESS_KEY_ID and BEDROCK_AWS_SECRET_ACCESS_KEY values must be configured or removed together in the shell and ~/.openwiki/.env.`.trim(),
-        );
-        return;
-      }
-
-      const apiKeyEnvKey = getProviderApiKeyEnvKey(currentProvider);
-
-      if (!apiKeyEnvKey) {
-        const hint = getProviderCredentialHint(currentProvider);
-
-        setError(
-          `${getProviderLabel(currentProvider)} does not use an API key.${
-            hint ? ` ${hint}` : ""
-          }`,
-        );
-        return;
-      }
-
-      setError(null);
-      setNotice(`Paste your ${getProviderLabel(currentProvider)} API key.`);
-      setSecretInputMode({
-        envKey: apiKeyEnvKey,
-        kind: "api-key",
-        label: `${getProviderLabel(currentProvider)} API key`,
-        provider: currentProvider,
-      });
-      setInputState({ cursorPosition: 0, value: "" });
-      setMenuState({ kind: "none" });
-      return;
-    }
-
-    if (option.id === "langsmith-key") {
-      if (args && args.length > 0) {
-        setError(
-          "Use the masked prompt for LangSmith keys; do not pass keys inline.",
-        );
-        return;
-      }
-
-      setError(null);
-      setNotice("Paste your LangSmith API key, or press Enter empty to clear.");
-      setSecretInputMode({
-        envKey: "LANGSMITH_API_KEY",
-        kind: "langsmith-key",
-        label: "LangSmith API key",
-      });
-      setInputState({ cursorPosition: 0, value: "" });
-      setMenuState({ kind: "none" });
-      return;
-    }
-
-    if (option.id === "init" || option.id === "update") {
-      resetInput();
-      onCommandRun(option.id, args);
-      return;
-    }
-
-    if (option.id === "clear") {
-      resetInput();
-      onClear();
-      setNotice("Started a new chat thread.");
-      return;
-    }
-
-    if (option.id === "help") {
-      resetInput();
-      setNotice(
-        "Slash commands: /provider, /model, /api-key, /langsmith-key, /init, /update, /clear, /help, /exit. Use arrows to select.",
-      );
-      return;
-    }
-
-    resetInput();
-    onSubmit("/exit");
-  }
-
-  async function selectModelMenuOption(selectedIndex: number) {
-    const option = getModelMenuOptions(currentModelId, currentProvider)[
-      selectedIndex
-    ];
-
-    if (!option) {
-      setError("Select a model.");
-      return;
-    }
-
-    if (option.kind === "custom") {
-      setError(null);
-      setNotice("Type a custom model ID after /model.");
-      setInputValue("/model ");
-      return;
-    }
-
-    await saveModelSelection(option.modelId);
-  }
-
-  async function saveModelSelection(rawModelId: string) {
-    const modelId = normalizeModelId(rawModelId);
-
-    if (!isValidModelId(modelId)) {
-      setError("Enter a valid model ID.");
-      return;
-    }
-
-    setIsSaving(true);
-    setError(null);
-    setNotice(null);
-
-    try {
-      await onModelSelect(modelId);
-      resetInput();
-      setNotice(`Model switched to ${modelId}.`);
-    } catch (saveError) {
-      setError(
-        saveError instanceof Error
-          ? saveError.message
-          : "Failed to save model selection.",
-      );
-    } finally {
-      setIsSaving(false);
-    }
-  }
-
-  async function selectProviderMenuOption(selectedIndex: number) {
-    const provider = SELECTABLE_OPENWIKI_PROVIDERS[selectedIndex];
-
-    if (!provider) {
-      setError("Select a provider.");
-      return;
-    }
-
-    await saveProviderSelection(provider);
-  }
-
-  async function saveProviderSelection(rawProvider: string) {
-    const provider = normalizeProvider(rawProvider);
-
-    if (provider === null) {
-      setError(
-        `Enter a valid provider: ${SELECTABLE_OPENWIKI_PROVIDERS.join(", ")}.`,
-      );
-      return;
-    }
-
-    setIsSaving(true);
-    setError(null);
-    setNotice(null);
-
-    try {
-      await onProviderSelect(provider);
-      resetInput();
-      const apiKeyEnvKey = getProviderApiKeyEnvKey(provider);
-      const requirement = providerUsesAwsSdkCredentials(provider)
-        ? (getProviderCredentialHint(provider) ??
-          "Configure AWS SDK credentials.")
-        : apiKeyEnvKey
-          ? `Ensure ${apiKeyEnvKey} is set.`
-          : `Ensure ${getProviderProjectEnvKey(provider)} is set. ${getProviderCredentialHint(provider) ?? ""}`.trim();
-      const modelNotice =
-        getProviderModelOptions(provider).length > 0
-          ? ` with model ${getDefaultModelId(provider)}`
-          : ". Set a model with /model";
-
-      setNotice(
-        `Provider switched to ${getProviderLabel(provider)}${modelNotice}. ${requirement}`,
-      );
-    } catch (saveError) {
-      setError(
-        saveError instanceof Error
-          ? saveError.message
-          : "Failed to save provider selection.",
-      );
-    } finally {
-      setIsSaving(false);
-    }
-  }
-
-  async function saveSecretInput() {
-    if (secretInputMode === null) {
-      return;
-    }
-
-    const nextValue = input.trim();
-    if (secretInputMode.kind === "api-key" && nextValue.length === 0) {
-      setError(`${secretInputMode.envKey} is required.`);
-      return;
-    }
-
-    setIsSaving(true);
-    setError(null);
-    setNotice(null);
-
-    try {
-      if (secretInputMode.kind === "langsmith-key") {
-        await saveOpenWikiEnv({
-          LANGCHAIN_PROJECT: nextValue.length > 0 ? "openwiki" : "",
-          LANGCHAIN_TRACING_V2: nextValue.length > 0 ? "true" : "false",
-          LANGSMITH_API_KEY: nextValue,
-        });
-      } else {
-        await saveOpenWikiEnv({
-          [secretInputMode.envKey]: nextValue,
-        });
-      }
-
-      const savedLabel = secretInputMode.label;
-      resetInput();
-      setSecretInputMode(null);
-      setNotice(`${savedLabel} saved.`);
-    } catch (saveError) {
-      setError(
-        saveError instanceof Error
-          ? saveError.message
-          : "Failed to save credential.",
-      );
-    } finally {
-      setIsSaving(false);
-    }
-  }
-
-  function resetInput() {
-    setInputState({ cursorPosition: 0, value: "" });
-    setMenuState({ kind: "none" });
-    setError(null);
-  }
-
-  function setInputValue(value: string) {
-    setInputState({
-      cursorPosition: value.length,
-      value,
-    });
-  }
-
-  const beforeCursor = input.slice(0, cursorPosition);
-  const afterCursor = input.slice(cursorPosition);
-
-  return (
-    <Box flexDirection="column" marginTop={1}>
-      <Box borderStyle="single" borderColor="blue" paddingX={1}>
-        <Text>
-          <Text color="blue">{">"}</Text>{" "}
-          {secretInputMode !== null ? (
-            <>
-              <Text color="gray">{secretInputMode.envKey}=</Text>
-              <Text color="yellow">{formatSecretInputSummary(input)}</Text>
-            </>
-          ) : input.length > 0 ? (
-            <>
-              {beforeCursor}
-              <InputCursor />
-              {afterCursor}
-            </>
-          ) : (
-            <>
-              <InputCursor />
-              <Text color="gray"> Ask a follow-up...</Text>
-            </>
-          )}
-        </Text>
-      </Box>
-      <Text>
-        <Text color="gray">
-          {secretInputMode !== null
-            ? "enter to save - esc to cancel - input is masked"
-            : `enter to send - / for commands - /exit to quit - cwd ${formatCwd(
-                process.cwd(),
-              )}`}
-        </Text>
-      </Text>
-      {secretInputMode !== null ? (
-        <Box flexDirection="column" marginTop={1}>
-          <Text color="gray">{secretInputMode.label}</Text>
-          <Text>
-            Saving to <Text color="cyan">{secretInputMode.envKey}</Text>
-          </Text>
-          {secretInputMode.kind === "langsmith-key" ? (
-            <Text color="gray">Press Enter empty to clear LangSmith.</Text>
-          ) : null}
-        </Box>
-      ) : menuState.kind !== "none" ? (
-        <SlashMenu
-          currentModelId={currentModelId}
-          currentProvider={currentProvider}
-          input={input}
-          menuState={menuState}
-        />
-      ) : null}
-      {notice ? <Text color="green">{notice}</Text> : null}
-      {isSaving ? <Text color="gray">Saving selection...</Text> : null}
-      {error ? <Text color="red">{error}</Text> : null}
-    </Box>
-  );
-}
-
-function SlashMenu({
-  currentModelId,
-  currentProvider,
-  input,
-  menuState,
-}: {
-  currentModelId: string;
-  currentProvider: OpenWikiProvider;
-  input: string;
-  menuState: Exclude<ChatInputMenuState, { kind: "none" }>;
-}) {
-  if (menuState.kind === "model") {
-    const modelOptions = getModelMenuOptions(currentModelId, currentProvider);
-
-    return (
-      <Box flexDirection="column" marginTop={1}>
-        <Text color="gray">Models for {getProviderLabel(currentProvider)}</Text>
-        {modelOptions.map((option, index) => (
-          <MenuRow
-            description={
-              option.kind === "model" && option.modelId === currentModelId
-                ? "current"
-                : option.kind === "custom"
-                  ? "type /model <model-id>"
-                  : ""
-            }
-            isSelected={index === menuState.selectedIndex}
-            key={option.label}
-            label={option.label}
-          />
-        ))}
-        {input.startsWith("/model ") ? (
-          <Text color="gray">Press enter to save the custom model ID.</Text>
-        ) : (
-          <Text color="gray">Use arrows, enter to select, esc to cancel.</Text>
-        )}
-      </Box>
-    );
-  }
-
-  if (menuState.kind === "provider") {
-    return (
-      <Box flexDirection="column" marginTop={1}>
-        <Text color="gray">Providers</Text>
-        {SELECTABLE_OPENWIKI_PROVIDERS.map((provider, index) => (
-          <MenuRow
-            description={
-              provider === currentProvider
-                ? "current"
-                : `default model ${getDefaultModelId(provider)}`
-            }
-            isSelected={index === menuState.selectedIndex}
-            key={provider}
-            label={getProviderLabel(provider)}
-          />
-        ))}
-        <Text color="gray">Use arrows, enter to select, esc to cancel.</Text>
-      </Box>
-    );
-  }
-
-  return (
-    <Box flexDirection="column" marginTop={1}>
-      <Text color="gray">Commands</Text>
-      {slashCommandOptions.map((option, index) => (
-        <MenuRow
-          description={option.description}
-          isSelected={index === menuState.selectedIndex}
-          key={option.id}
-          label={option.label}
-        />
-      ))}
-      <Text color="gray">Use arrows, enter to select, esc to cancel.</Text>
-    </Box>
-  );
-}
-
-function MenuRow({
-  description,
-  isSelected,
-  label,
-}: {
-  description: string;
-  isSelected: boolean;
-  label: string;
-}) {
-  return (
-    <Text>
-      <Text color={isSelected ? "cyan" : "gray"}>{isSelected ? ">" : " "}</Text>{" "}
-      <Text bold={isSelected}>{label.padEnd(28)}</Text>
-      <Text color="gray">{description}</Text>
-    </Text>
-  );
-}
-
-function InputCursor() {
-  return <Text color="cyan">|</Text>;
-}
-
-function PromptBlock({ message }: { message: string }) {
-  return (
-    <Box flexDirection="column" marginBottom={1}>
-      <Text backgroundColor="gray" wrap="wrap">
-        {" "}
-        <Text color="cyan">{">"}</Text> {message}
-      </Text>
-    </Box>
-  );
-}
-
 function updateRunningCredentialDiagnostics(
   state: RunState,
   credentialDiagnostics: CredentialDiagnostic[],
@@ -2444,46 +948,6 @@ function updateRunningCredentialDiagnostics(
         credentialDiagnostics,
       }
     : state;
-}
-
-type PanelProps = {
-  title: string;
-  children: React.ReactNode;
-};
-
-function Panel({ title, children }: PanelProps) {
-  return (
-    <Box flexDirection="column" marginBottom={1}>
-      <Text>
-        <Text color="cyan"># </Text>
-        <Text bold>{title}</Text>
-      </Text>
-      <Box flexDirection="column" marginLeft={2}>
-        {children}
-      </Box>
-    </Box>
-  );
-}
-
-type RowsProps = {
-  rows: HelpRow[];
-};
-
-function Rows({ rows }: RowsProps) {
-  const labelWidth = Math.max(...rows.map((row) => row.label.length));
-
-  return (
-    <>
-      {rows.map((row) => (
-        <Text key={row.label}>
-          {"  "}
-          {row.label.padEnd(labelWidth)}
-          {"  "}
-          {row.description}
-        </Text>
-      ))}
-    </>
-  );
 }
 
 const argv = process.argv.slice(2);
